HomeViewModel and HomeView Refactor
