+++ Fitness_App/Activity.swift
@@ -1,0 +1,21 @@
+//
+//  Activity.swift
+//  Fitness_App
+//
+//  Created by Snehal Chavan on 4/20/25.
+//
+
+import Foundation
+
+import SwiftUI
+
+struct Activity{
+    
+    let id: Int
+    let title: String
+    let subtitle: String
+    let image: String
+    let tintColor: Color
+    let amount: String
+    
+}
