+++ Fitness_App/ActivityCardView.swift
@@ -9,16 +9,7 @@ import SwiftUI
 
 // model for activity to make it dynamic
 
-struct Activity{
-    
-    let id: Int
-    let title: String
-    let subtitle: String
-    let image: String
-    let tintColor: Color
-    let amount: String
-    
-}
+
 
 struct ActivityCardView: View {
