+++ Fitness_App/HomeView.swift
@@ -7,38 +7,17 @@
 
 import SwiftUI
 
+
+
 struct HomeView: View {
     
-    @State var calories: Int = 123
-    @State var active: Int = 123
-    @State var stand: Int = 8
+    @StateObject var viewModel = HomeViewModel()
+    
+    
     
     // Mockdata
     
-    var MockActivity = [
-        
-        Activity(id: 0, title: "Today Steps", subtitle: " Goal 35000", image: "figure.walk", tintColor: .green, amount: "9800"),
-        
-        Activity(id: 1, title: "Today Steps", subtitle: " Goal 54009", image: "figure.walk", tintColor: .blue, amount: "6979"),
-        
-        Activity(id: 2, title: "Today Steps", subtitle: " Goal 64321", image: "figure.walk", tintColor: .red, amount: "4500"),
-        
-        Activity(id: 3, title: "Today Steps", subtitle: " Goal 55533", image: "figure.run", tintColor: .yellow, amount: "59800")
-        
-        
-    ]
-    
-    var mockWorkouts = [
-        
-        Workout(id: 0, title: "Running", image: "figure.run", tintColor: .cyan, duration: "45 mins", date: "April 14", calories: "520 kcal"),
-        
-        Workout(id: 1, title: "Strength Training", image: "figure.run", tintColor: .green, duration: "25 mins", date: "April 14", calories: "150 kcal"),
-        
-        Workout(id: 2, title: "Pilates", image: "figure.walk", tintColor: .blue, duration: "30 mins", date: "April 14", calories: "300 kcal"),
-        
-        Workout(id: 3, title: "Legs", image: "figure.run", tintColor: .red, duration: "37 mins", date: "April 14", calories: "600 kcal")
-    ]
-    
+   
     
     var body: some View {
         
@@ -57,7 +36,7 @@ struct HomeView: View {
                     HStack{
                         Spacer() // pushes everything on the left of screen
                         
-                        VStack{
+                        VStack(alignment: .leading){
                             
                             VStack(alignment: .leading, spacing: 8){
                                 
@@ -66,7 +45,7 @@ struct HomeView: View {
                                     .bold()
                                     .foregroundStyle(.red)
                                 
-                                Text("123 kcal")
+                                Text("\(viewModel.calories)")
                                     .bold()
                             }
                             .padding(.bottom)
@@ -78,7 +57,7 @@ struct HomeView: View {
                                     .bold()
                                     .foregroundStyle(.green)
                                 
-                                Text("123 kcal")
+                                Text("\(viewModel.active)")
                                     .bold()
                             }
                             .padding(.bottom)
@@ -90,7 +69,7 @@ struct HomeView: View {
                                     .bold()
                                     .foregroundStyle(.red)
                                 
-                                Text("8 hours")
+                                Text("\(viewModel.stand)")
                                     .bold()
                             }
                             .padding(.bottom)
@@ -103,13 +82,13 @@ struct HomeView: View {
                         ZStack{
                             
                             // Calling progressCircleView
-                            ProgressCircleView(progress: $calories, color: .red, goal: 600)
+                            ProgressCircleView(progress: $viewModel.calories, color: .red, goal: 600)
                             
                             
-                            ProgressCircleView(progress: $calories, color: .green, goal: 900)
+                            ProgressCircleView(progress: $viewModel.calories, color: .green, goal: 900)
                                 .padding(20)
                             
-                            ProgressCircleView(progress: $calories, color: .blue, goal: 600)
+                            ProgressCircleView(progress: $viewModel.calories, color: .blue, goal: 600)
                                 .padding(.all, 40)
                             
                         }
@@ -144,7 +123,7 @@ struct HomeView: View {
                     // uploads pieces of data at a time to avoid bottleneck
                     LazyVGrid(columns: Array(repeating: GridItem(spacing: 20), count: 2))
                     {
-                        ForEach(MockActivity, id: \.id ){
+                        ForEach(viewModel.MockActivity, id: \.id ){
                             activity in ActivityCardView(activity: activity)
                         }
                     }
@@ -174,7 +153,7 @@ struct HomeView: View {
                     // uploads pieces of data at a time to avoid bottleneck
                     LazyVStack{
                         
-                        ForEach(mockWorkouts, id: \.id ){
+                        ForEach(viewModel.mockWorkouts, id: \.id ){
                             workout in WorkoutCardView(workout: workout)
                         }
                     }
