+++ Fitness_App/Workout.swift
@@ -1,0 +1,22 @@
+//
+//  Workout.swift
+//  Fitness_App
+//
+//  Created by Snehal Chavan on 4/20/25.
+//
+
+import Foundation
+import SwiftUI
+
+struct Workout{
+    
+    let id: Int
+    let title: String
+    let image: String
+    let tintColor: Color
+    let duration: String
+    let date: String
+    let calories: String
+    
+    
+}
