+++ Fitness_App/WorkoutCardView.swift
@@ -7,18 +7,7 @@
 
 import SwiftUI
 
-struct Workout{
-    
-    let id: Int
-    let title: String
-    let image: String
-    let tintColor: Color
-    let duration: String
-    let date: String
-    let calories: String
-    
-    
-}
+
 
 struct WorkoutCardView: View {
